backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
python-dotenv
aiofiles
httpx
supabase
groq
pdfplumber
pytesseract
PyPDF2
Pillow
